Add keybinding config file
This will add a `config.yml` file along side the `client.yml` file which
will allow the user to configure custom keybindings.  Currently this
commit only covers global keybindings and not per-screen keybindings but
this could be added in the future. This commit also doesn't currently
allow for overriding default navigation keys (hjkl etc.) but that could
also be added later.

Resolves: #46