Get rid of EpisodeTableContext
There is currently no such thing as a context for playing episodes, and
the only collection of episodes that is possible is the episodes from a
show, so there's not currently a need to handle the playing or
navigation context of the episode table. If there is some time in the
future when it makes sense, it can be implemented then.