Don't add analysis view to stack if already in it.
Before this commit, if a user pressed `v` multiple times, they would
have to press `q` the same number of times to get to the previous view.

This commit makes it so that an analysis view entry is only pushed into
the navigation stack if the current view isn't an analysis one.