Implement paging key handlers
This commit implements the page fetching logic for pageable lists. In
summary, the logic is to always try to have enough items in the list
ahead that we can jump a full UI page into the list. To do this, we
check the state of the list and its current index when the list is drawn
to see if wee need to fetch another page.

This requires synchronization between the pageable list and the network
IO handler since we could easily requst the same page multiple times if
we don't lock the operation. The caveat here is that each network event
handler (for paging) will have to check the lock when it is done and set
it to false to let thelist know that is safe to askfor more pages again.