Reformat help menu table to have one column (#533)
If the terminal width is small, then the constraints on table column
widths cannot be satisfied. In this case, the Cassowary solver chooses a
constraint to be violated non-deterministically, which causes a
flickering rendering.

This workaround formats each row of the table into a single string,
creating a single-column table. With one column and a "Max" constraint
on its width, the non-determinism is eliminated.