feat: Allow wayland builds without xcb dependency
The arbout carte is used to interact with the system clipboard. This
however imposes a dependency on xcb on the crate for Linux/Unix builds.

On a pure Wayland system this is undesirable as interaction with the
system clipboard via xcb in only possible if the user runs a xwayland
server. This cause spotify-cli to require a dependency to a x11 c lib
without providing any functionally to the user.

This change intoduces to new cargo features to spotify-cli: `x11` and
`wayland`.

The `x11` feature is in the default feature and will preserve the
current application behavior.

The `wayland` feature will disable the dependency onto arboard and use
`#[cfg(feature = "wayland")]` attributes to disable code using this
crate, as well as any reference from the help menu to any copy
functionally.